+++ packages/storage/src/index.ts
@@ -21,9 +21,20 @@ import {
   createLocalStorage,
   createSessionStorage,
 } from "./storage.js";
-import { CookieOptions, cookieStorage, createCookieStorage } from "./cookies.js";
-import { addClearMethod } from "./tools.js";
-import { PersistenceOptions, makePersisted } from "./persisted.js";
+import { type CookieOptions, cookieStorage, createCookieStorage } from "./cookies.js";
+import { addClearMethod, addWithOptionsMethod, multiplexStorage } from "./tools.js";
+import {
+  type PersistenceOptions,
+  type PersistenceSyncAPI,
+  type PersistenceSyncData,
+  type PersistenceBaseOptions,
+  type PersistenceSyncCallback,
+  makePersisted,
+  multiplexSync,
+  storageSync,
+  messageSync,
+  wsSync,
+} from "./persisted.js";
 export {
   createStorage,
   createAsyncStorage,
@@ -34,6 +45,16 @@ export {
   type CookieOptions,
   cookieStorage,
   addClearMethod,
+  addWithOptionsMethod,
+  multiplexStorage,
+  type PersistenceBaseOptions,
   type PersistenceOptions,
+  type PersistenceSyncCallback,
+  type PersistenceSyncData,
+  type PersistenceSyncAPI,
   makePersisted,
+  multiplexSync,
+  storageSync,
+  wsSync,
+  messageSync,
 };